drivers/common_psql: share more code
