+++ datacube/drivers/postgres/_core.py
@@ -22,10 +22,10 @@ from datacube.drivers.common_psql import (
     transfer_ownership,
     transfers_required,
 )
+from datacube.drivers.common_psql.sql import INSTALL_TRIGGER_SQL_TEMPLATE
 from datacube.drivers.postgres.sql import (
     ADDED_COLUMN_INDEX_SQL_TEMPLATE,
     ADDED_COLUMN_MIGRATE_SQL_TEMPLATE,
-    INSTALL_TRIGGER_SQL_TEMPLATE,
     SCHEMA_NAME,
     TYPES_INIT_SQL,
     UPDATE_COLUMN_INDEX_SQL_TEMPLATE,
