+++ datacube/drivers/postgres/sql.py
@@ -6,31 +6,23 @@
 Custom types for postgres & sqlalchemy
 """
 
-import warnings
-
-from sqlalchemy import TIMESTAMP, Connection, inspect, text
+from sqlalchemy import TIMESTAMP, Connection
 from sqlalchemy.dialects.postgresql.ranges import AbstractRange, Range
 from sqlalchemy.ext.compiler import compiles
-from sqlalchemy.sql import sqltypes
-from sqlalchemy.sql.expression import ClauseElement, Executable
 from sqlalchemy.sql.functions import GenericFunction
 from sqlalchemy.types import Double
+
+import datacube.drivers.common_psql.sql as common_psql
 
 SCHEMA_NAME = "agdc"
 
 
-class CreateView(Executable, ClauseElement):
-    inherit_cache = True
-
-    def __init__(self, name: str, select) -> None:
-        self.name = name
-        self.select = select
-
-
-@compiles(CreateView)
-def visit_create_view(element, compiler, **kw) -> str:
-    return f"CREATE VIEW {element.name} AS {compiler.process(element.select, literal_binds=True)}"
-
+# Make old names available from this module.
+from datacube.drivers.common_psql.sql import (  # noqa: E402, F401
+    INSTALL_TRIGGER_SQL_TEMPLATE,
+    CreateView,
+    visit_create_view,
+)
 
 UPDATE_TIMESTAMP_SQL: str = f"""
 create or replace function {SCHEMA_NAME}.set_row_update_time()
@@ -61,16 +53,6 @@ ADDED_COLUMN_INDEX_SQL_TEMPLATE = """
 create index if not exists ix_{table}_added
 on {schema}.{table}(added);
 """
-
-INSTALL_TRIGGER_SQL_TEMPLATE = [
-    "drop trigger if exists row_update_time_{table} on {schema}.{table}",
-    """
-    create trigger row_update_time_{table}
-    before update on {schema}.{table}
-    for each row
-    execute procedure {schema}.set_row_update_time();
-    """,
-]
 
 TYPES_INIT_SQL: list[str] = [
     f"""
@@ -127,22 +109,12 @@ class Float8Range(GenericFunction):
         self.packagenames = (f"{SCHEMA_NAME}",)
 
 
-class PGNAME(sqltypes.Text):
-    """Postgres 'NAME' type."""
-
-    __visit_name__ = "NAME"
-
-
-@compiles(PGNAME)
-def visit_name(element, compiler, **kw) -> str:
-    return "NAME"
-
-
-def pg_exists(conn, name: str) -> bool:
-    """
-    Does a postgres object exist?
-    """
-    return conn.execute(text(f"SELECT to_regclass('{name}')")).scalar() is not None
+# Make old names available from this module.
+from datacube.drivers.common_psql.sql import (  # noqa: E402, F401
+    PGNAME,
+    pg_exists,
+    visit_name,
+)
 
 
 def pg_column_exists(
@@ -151,10 +123,4 @@ def pg_column_exists(
     """
     Does a table column exist?
     """
-    if table.startswith((f"{SCHEMA_NAME}.", f"'{SCHEMA_NAME}.", f'"{SCHEMA_NAME}.')):
-        warnings.warn(
-            f"Call pg_column_exists with a table name without {SCHEMA_NAME}.",
-            stacklevel=2,
-        )
-        table = table.replace(f"{SCHEMA_NAME}.", "")
-    return column in [x.get("name") for x in inspect(conn).get_columns(table, schema)]
+    return common_psql.pg_column_exists(conn, table, column, schema, SCHEMA_NAME)
